+++ src/hotspot/share/services/mallocHeader.hpp
@@ -108,7 +108,9 @@ class MallocHeader {
 
   void print_block_on_error(outputStream* st, address bad_address) const;
 
-  static uint16_t build_footer(uint8_t b1, uint8_t b2) { return ((uint16_t)b1 << 8) | (uint16_t)b2; }
+  static uint16_t build_footer(uint8_t b1, uint8_t b2) {
+    return static_cast<uint16_t>(((uint16_t)b1 << 8) | (uint16_t)b2);
+  }
 
   uint8_t* footer_address() const   { return ((address)this) + sizeof(MallocHeader) + _size; }
   uint16_t get_footer() const       { return build_footer(footer_address()[0], footer_address()[1]); }
